Move MFA related types from otp to auth
