+++ src/components/Users/TwoFactorAuth.tsx
@@ -15,8 +15,8 @@ import { TOTPSetupDialog } from "@/components/Users/TOTPSetupDialog";
 
 import mutate from "@/Utils/request/mutate";
 import { HTTPError, StructuredError } from "@/Utils/request/types";
+import { BackupCodesRespone, TOTPSetupResponse } from "@/types/auth/auth";
 import authApi from "@/types/auth/authApi";
-import { TOTPSetupResponse, TOTPVerifyResponse } from "@/types/auth/otp";
 
 interface DialogState {
   password: boolean;
@@ -93,7 +93,7 @@ export const TwoFactorAuth = ({ userData }: userChildProps) => {
 
   const { mutate: verifyTOTP, isPending: isVerifying } = useMutation({
     mutationFn: mutate(authApi.totp.verify),
-    onSuccess: (data: TOTPVerifyResponse) => {
+    onSuccess: (data: BackupCodesRespone) => {
       if (data.backup_codes && Array.isArray(data.backup_codes)) {
         setBackupCodes(data.backup_codes);
         closeSetupDialog();
@@ -130,7 +130,7 @@ export const TwoFactorAuth = ({ userData }: userChildProps) => {
   const { mutate: regenerateBackupCodes, isPending: isRegenerating } =
     useMutation({
       mutationFn: mutate(authApi.totp.regenerateBackupCodes),
-      onSuccess: (data: { backup_codes: string[] }) => {
+      onSuccess: (data: BackupCodesRespone) => {
         setBackupCodes(data.backup_codes);
         closeRegenerateConfirm();
         setShowRegenerateBackupCodes(true);
